app/requirements.txt
# ARVTON Backend — Python Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
slowapi>=0.1.9
python-dotenv>=1.0.0

# ML / Pipeline
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
Pillow>=10.2.0
opencv-python-headless>=4.9.0
trimesh>=4.1.0
transformers>=4.38.0
diffusers>=0.27.0
safetensors>=0.4.2
huggingface-hub>=0.21.0

# S3 (optional)
# boto3>=1.34.0
